app/python-service/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
base58==2.1.1
boto3==1.35.10
botocore==1.35.10
cachetools==4.2.4
certifi==2024.7.4
charset-normalizer==3.3.2
construct==2.10.68
construct-typing==0.5.6
exceptiongroup==1.2.2
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.8
jmespath==1.0.1
jsonalias==0.1.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
solana==0.30.0
solders==0.15.1
tqdm==4.66.5
types-cachetools==4.2.10
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
websockets==10.4
